refactor: Hibernate configs and testing

+++ src/main/java/utils/DB.java
@@ -3,17 +3,17 @@ package utils;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
-
 import org.hibernate.Session;
-
 import tukano.api.Result;
-import tukano.api.User;
-import tukano.api.Short;
 import tukano.api.Result.ErrorCode;
-import tukano.impl.storage.azure.CosmosDB;
+import tukano.api.Short;
+import tukano.api.User;
 import tukano.impl.cache.RedisCache;
+import tukano.impl.storage.azure.CosmosDB;
 
 public class DB {
+
+    private static final boolean useHibernate = true; // TODO for testing purposes
 
     public static <T> List<T> sql(String query, Class<T> clazz) {
         return Hibernate.getInstance().sql(query, clazz);
@@ -25,20 +25,24 @@ public class DB {
 
     public static <T> Result<T> getOne(String id, Class<T> clazz) {
         String[] classSplit = clazz.toString().toLowerCase().split("\\.");
-        String className = classSplit[classSplit.length-1];
+        String className = classSplit[classSplit.length - 1];
         // TODO this is kinda disgusting but avoids type checking so..? lmk
 
         String key = className + ":" + id;
         var res = RedisCache.getOne(key, clazz);
 
         if (res.isOK()) {
-            System.out.println("Cached result: " + res.getClass() + " | " + res);
+            System.out.println("Cached result: " + res.getClass() + " | " +
+                               res);
             return res;
         }
 
         System.out.println("Cache miss: " + key);
-        // return CosmosDB.getInstance().getOne(id, clazz);
-        return Hibernate.getInstance().getOne(id, clazz);
+
+        if (useHibernate) {
+            return Hibernate.getInstance().getOne(id, clazz);
+        }
+        return CosmosDB.getInstance().getOne(id, clazz);
     }
 
     public static <T> Result<T> deleteOne(T obj) {
@@ -46,8 +50,10 @@ public class DB {
         if (key != null)
             RedisCache.invalidate(key);
 
-        // return Result.errorOrValue(CosmosDB.getInstance().deleteOne(obj), obj);
-        return Hibernate.getInstance().deleteOne(obj);
+        if (useHibernate) {
+            return Hibernate.getInstance().deleteOne(obj);
+        }
+        return Result.errorOrValue(CosmosDB.getInstance().deleteOne(obj), obj);
     }
 
     public static <T> Result<T> updateOne(T obj) {
@@ -55,8 +61,10 @@ public class DB {
         if (key != null)
             RedisCache.invalidate(key);
 
-        // return CosmosDB.getInstance().updateOne(obj);
-        return Hibernate.getInstance().updateOne(obj);
+        if (useHibernate) {
+            return Hibernate.getInstance().updateOne(obj);
+        }
+        return CosmosDB.getInstance().updateOne(obj);
     }
 
     public static <T> Result<T> insertOne(T obj) {
@@ -66,8 +74,11 @@ public class DB {
         if (key != null)
             RedisCache.insertOne(key, obj);
 
-        // return Result.errorOrValue(CosmosDB.getInstance().insertOne(obj), obj);
-        return Result.errorOrValue(Hibernate.getInstance().persistOne(obj), obj);
+        if (useHibernate) {
+            return Result.errorOrValue(Hibernate.getInstance().persistOne(obj),
+                                       obj);
+        }
+        return Result.errorOrValue(CosmosDB.getInstance().insertOne(obj), obj);
     }
 
     public static <T> Result<T> transaction(Consumer<Session> c) {
